requirements.txt
langchain
langchain-community
langchain-openai
langchain-pinecone
pinecone-client
sentence-transformers
pypdf
python-dotenv
streamlit